+++ packages/types/custom/core.ts
@@ -375,6 +375,7 @@ export interface GfiConfiguration extends PluginOptions {
   layers: Record<string, GfiLayerConfiguration>
   activeLayerPath?: string
   afterLoadFunction?: GfiAfterLoadFunction
+  boxSelect?: boolean
   /**
    * If required the stroke and fill of the highlighted feature can be configured.
    * Otherwise, a default style is applied.
@@ -394,7 +395,7 @@ export interface GfiConfiguration extends PluginOptions {
    */
   maxFeatures?: number
   mode?: 'bboxDot' | 'intersects'
-  boxSelect?: boolean
+  multiSelect?: 'box' | 'circle'
   renderType?: RenderType
 }
 
@@ -416,11 +417,6 @@ export interface LegendConfiguration extends PluginOptions {
     open?: string
     close?: string
   }
-}
-
-export interface MeasureConfiguration extends PluginOptions {
-  color?: Color | ColorLike
-  textColor?: Color | ColorLike
 }
 
 export interface AppearOnClick {
@@ -668,7 +664,6 @@ export interface MapConfig extends MasterportalApiConfig {
   gfi?: GfiConfiguration
   iconMenu?: IconMenuConfiguration
   legend?: LegendConfiguration
-  measure?: MeasureConfiguration
   pins?: PinsConfiguration
   reverseGeocoder?: ReverseGeocoderConfiguration
   scale?: ScaleConfiguration