+++ packages/plugins/Pins/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+# unpublished
+
+- Feature: Recognize the multi-select-`Draw` of `@polar/plugin-gfi`.
+
 ## 2.0.0
 
 - Breaking: Upgrade peerDependency `ol` from `^7.1.0` to `^9.2.4`.